+++ test/data/generate_schedule_csvs.jl
@@ -2,8 +2,8 @@ using DelimitedFiles
 
 # add a factor of 2x and sign inversion based on D-Wave Hamiltonian convention
 # https://docs.dwavesys.com/docs/latest/c_qpu_annealing.html
-A = (s) -> -2.0*cos(π/2*s)
-B = (s) ->  2.0*sin(π/2*s)
+A = (s) -> -2.0*cos(π/2*s)/(2.0*π)
+B = (s) ->  2.0*sin(π/2*s)/(2.0*π)
 
 header = ["s", "A(s)", "B(s)"]