+++ CHANGELOG.md
@@ -5,6 +5,7 @@ QuantumAnnealing.jl Change Log
 - Add variant of `solve_de` with adaptive solve tolerance
 - Add tools for working with classical Ising models (#8)
 - Add data processing tools (#6)
+- Fix units in dwave annealing schedules (#17)
 
 ### v0.0.1
 - Initial release